requirements.txt
streamlit==1.32.2
pandas==2.2.1
scikit-learn==1.4.1.post1
joblib==1.3.2
langgraph==0.0.30
langchain==0.1.13
langchain-community==0.0.29
langchain-groq
langchain-huggingface
chromadb==0.4.24
python-dotenv==1.0.1
